Initial commit. Finds all words on board.
This commit marks the point before using qt to alter the code to have a
gui. The current state of the project is that we can do most things for
the game except actually playing the game i.e. we can initialize the
board and find all the words on the board relative to a given dictionary
of words.